backend/app/requirements.txt
uvicorn
fastapi
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
oss2
ffmpeg-python
